refactor: import components directly
